code/web_scraping/requirements.txt
requests 
beautifulsoup4 
readability-lxml 
pandas
google-genai 
python-dotenv